+++ serialno.h
@@ -5,10 +5,11 @@
 //
 // Copyright (c) 2010 David Kirkby dkirkby@uci.edu
 // =====================================================================
+#include <stdint.h> // for uint32_t
 
 #define HUB_SERIAL_NUMBER_MASK 0xff000000 // all hubs have these bits set in their SN
 #define IS_HUB(sn) (((sn) & HUB_SERIAL_NUMBER_MASK) == HUB_SERIAL_NUMBER_MASK)
 
-extern unsigned long serialNumber();
+extern uint32_t serialNumber();
 
 #endif
